refactor: move CoderFromUrn into worker mod
